+++ src/ObjectB/processing.py
@@ -1,13 +1,8 @@
-from scipy.optimize import curve_fit
-import math
-import xlsxwriter
-from django.contrib.staticfiles.storage import staticfiles_storage
 import os
 from django.conf import settings
 import base64
 from io import BytesIO
 import pandas as pd
-import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib
 matplotlib.use('Agg')
@@ -43,9 +38,6 @@ def getZ(perc):
     for i in range(len(val_list)):
         for j in range(len(val_list[i])):
             if val_list[i][j] > val:
-                # print(val)
-                # print("Row:", i, ", Column:", j)
-
                 z_val = (i / 10) + (j / 100)
                 return z_val
 
@@ -57,7 +49,6 @@ def value_verify():
     gas_mid = [float(GasPrice)] * 12
 
     create_table(oil_mid, gas_mid)
-    # create_gas_table(gas_mid)
 
 
 def create_table(oil_mid, gas_mid):
@@ -66,16 +57,16 @@ def create_table(oil_mid, gas_mid):
     oil_perc = [OilPrice]
     gas_perc = [GasPrice]
 
-    z_val = getZ(int(PercLine))
+    z_val = getZ(float(PercLine))
 
-    std_dev_oil_price = (int(OilPrice) * int(OilSD)) / 100
-    std_dev_gas_price = (float(GasPrice) * int(GasSD)) / 100
+    std_dev_oil_price = (float(OilPrice) * float(OilSD)) / 100
+    std_dev_gas_price = (float(GasPrice) * float(GasSD)) / 100
 
     oil_val = std_dev_oil_price
     gas_val = std_dev_gas_price
 
     for i in range(1, len(oil_mid)+1):
-        if int(PercLine) > 50:
+        if float(PercLine) > 50:
             if float(OilPrice) - ((i ** 0.5) * z_val * oil_val) > 0:
                 oil_perc.append(float(OilPrice) - ((i ** 0.5) * z_val * oil_val))
             else:
@@ -84,7 +75,7 @@ def create_table(oil_mid, gas_mid):
                 gas_perc.append(float(GasPrice) - ((i ** 0.5) * z_val * gas_val))
             else:
                 gas_perc.append(0)
-        elif int(PercLine) < 50:
+        elif float(PercLine) < 50:
             if float(OilPrice) + ((i ** 0.5) * z_val * oil_val) > 0:
                 oil_perc.append(float(OilPrice) + ((i ** 0.5) * z_val * oil_val))
             else:
@@ -121,21 +112,16 @@ def bar_plot(oil_perc, gas_perc):
     ax.set_title("Product Price Analysis, " + str(PercLine) + "% Probability of Exceeding the Green Line", fontsize=28)
 
     label = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-    # oil_perc = [int(a) for a in oil_perc]
 
     ax.scatter(label, oil_perc, color="orange", marker=".", s=250, linewidth=3)
     ax.set_xlabel("Time (Months)", fontsize=25)
 
     ax.plot(label, oil_perc)
-    # ax.plot(label, oil_perc, '*')
-
     plt.xlabel("Time (Months)")
-    # plt.ylim(min(oil_perc) - 5, max(oil_perc) + 5)
     ax.set_ylabel("Oil Price ($/BO)", fontsize=25)
     plt.xticks(fontsize=20)
     plt.yticks(fontsize=20)
     ax.set_title(str(PercLine) + "% Likelihood of Oil Price Being Above (months into the future)", fontsize=28)
-    # plt.show()
 
     buffer = BytesIO()
     plt.savefig(buffer, format='png')
@@ -149,26 +135,19 @@ def bar_plot(oil_perc, gas_perc):
     ChartOil = graph
 
     # -------------------- Gas Plot --------------------
-    # plt.close()
     fig, ax = plt.subplots(1, figsize=(16, 16))
     ax.set_title("Product Price Analysis, " + str(PercLine) + "% Probability of Exceeding the Green Line", fontsize=28)
 
     label = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-    # gas_perc = [int(a) for a in gas_perc]
-
     ax.scatter(label, gas_perc, color="orange", marker=".", s=250, linewidth=3)
     ax.set_xlabel("Time (Months)", fontsize=25)
 
     ax.plot(label, gas_perc)
-    # ax.plot(label, gas_perc, '*')
-
     plt.xlabel("Time (Months)")
-    # plt.ylim(min(gas_perc) - 5, max(gas_perc) + 5)
     ax.set_ylabel("Gas Price ($/MSCF)", fontsize=25)
     plt.xticks(fontsize=20)
     plt.yticks(fontsize=20)
     ax.set_title(str(PercLine) + "% Likelihood of Gas Price Being Above (months into the future)", fontsize=28)
-    # plt.show()
 
     buffer = BytesIO()
     plt.savefig(buffer, format='png')
